+++ components/__init__.py
@@ -5,3 +5,4 @@ from .alignment_lyrics_handler import AlignmentLyricsHandler
 from .audio_lyric_align_task import AudioLyricAlignTask
 from .lyric_sanitizer import LyricSanitizer
 from .text_simplifier import simplify
+from .lyric_state import LyricValidity